feat/tramplin-018 Перенес общие сущности в commons Взял security config'и из профиля (надо будет перенастроить под текущий сервис)

+++ server/commons/src/main/kotlin/ru/itplanet/trampline/commons/dao/dto/LocationDto.kt
@@ -31,10 +31,10 @@ open class LocationDto {
     open var postalCode: String? = null
 
     @Column(name = "latitude", precision = 9, scale = 6)
-    open var latitude: BigDecimal? = null
+    open var latitude: Double? = null
 
     @Column(name = "longitude", precision = 9, scale = 6)
-    open var longitude: BigDecimal? = null
+    open var longitude: Double? = null
 
     @Column(name = "is_active", nullable = false)
     open var isActive: Boolean = true
@@ -55,8 +55,8 @@ open class LocationDto {
         title: String? = null,
         addressLine2: String? = null,
         postalCode: String? = null,
-        latitude: BigDecimal? = null,
-        longitude: BigDecimal? = null
+        latitude: Double? = null,
+        longitude: Double? = null
     ) {
         this.city = city
         this.addressLine = addressLine
